move ParseResult model to models
