+++ src/aa_pbs_exporter/pbs_2022_01/models/parse_result.py
@@ -1,14 +1,8 @@
-# from dataclasses import dataclass
 import json
 
 from pydantic import BaseModel
 
 from aa_pbs_exporter.pbs_2022_01.models import raw
-
-# @dataclass
-# class ParseResult:
-#     current_state: str
-#     parsed_data: raw.ParsedIndexedString
 
 
 class ParseResult(BaseModel):
