+++ src/aa_pbs_exporter/pbs_2022_01/translate/result_handler.py
@@ -3,7 +3,7 @@ from uuid import uuid5
 from aa_pbs_exporter.pbs_2022_01 import PARSER_DNS
 from aa_pbs_exporter.pbs_2022_01.models import raw
 from aa_pbs_exporter.pbs_2022_01.models.common import HashedFile
-from aa_pbs_exporter.pbs_2022_01.parse_result import ParseResult
+from aa_pbs_exporter.pbs_2022_01.models.parse_result import ParseResult
 from aa_pbs_exporter.pbs_2022_01.raw_helpers import collect_calendar_entries
 from aa_pbs_exporter.pbs_2022_01.validate_raw import RawValidator
 from aa_pbs_exporter.snippets.indexed_string.state_parser.result_handler import (
